HW3/requirements.txt
torch
torchvision
tqdm
pillow
pytorch-fid